requirements.txt
you-get==0.4.1475
scikit-learn==0.24.1
paddlehub==1.8.3
paddlepaddle==2.0.1
paddleslim==1.1.1
matplotlib==3.3.4
flask==1.1.1
flask-apscheduler==1.12.1
flask-babel==2.0.0
requests==2.25.1
opencv-python==4.2.0.32
numpy==1.19.2
tensorflow-gpu==1.14.0
tensorflow-estimator==1.14.0
gensim==3.8.2
 JsonForm==0.0.2
jsonschema==2.6.0
JsonSir==0.0.2
six==1.11.0
jieba==0.42.1
sklearn==0.0
pyopenssl==17.0.0
joblib==0.14.1
pandas==0.23.0
parsel==1.2.0

pip install smart-open==1.10.0 -i https://pypi.tuna.tsinghua.edu.cn/simple
pip install paddlepaddle -i https://mirror.baidu.com/pypi/simple
matplotlib==3.3.4 -i https://pypi.tuna.tsinghua.edu.cn/simple
pip install opencv-python==4.2.0.32 -i https://pypi.tuna.tsinghua.edu.cn/simple
pip install paddlenlp==2.5.2 -i https://mirror.baidu.com/pypi/simple

pip install --upgrade pip
pyclipper tools
